# backend/openapi.yaml
openapi: 3.0.3
info:
  title: DinnerWise API
  version: 1.0.0
  description: OpenAPI spec for the DinnerWise backend.
...
paths:
...
  /events:
    get:
      tags: [Events]
      summary: Get all events
      parameters:
        - name: category_ids
          in: query
          schema: { type: array, items: { type: integer } }
        - name: restaurant_id
          in: query
          schema: { type: integer }
        - name: city_ids
          in: query
          schema: { type: array, items: { type: integer } }
        - name: price_ids
          in: query
          schema: { type: array, items: { type: integer } }
        - name: date
          in: query
          schema: { type: string }
        - name: order
          in: query
          schema: { type: string, enum: [price, availability] }
        - name: tag_ids
          in: query
          schema: { type: array, items: { type: integer } }
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
...
  /restaurants:
    get:
      tags: [Restaurants]
      summary: Get all restaurants
      parameters:
        - name: city
          in: query
          schema: { type: string }
      responses:
        '200':
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Restaurant' }
  /restaurants/search:
    get:
      tags: [Restaurants]
      summary: Search restaurants
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Restaurant' }
...
  /tags/search:
    get:
      tags: [Tags]
      summary: Search tags
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Tag' }
...
  /cities:
    get:
      tags: [Cities]
      summary: Get all cities
      parameters:
        - name: q
          in: query
          schema: { type: string }
      responses:
        '200':
          description: List of cities
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/City' }
  /cities/search:
    get:
      tags: [Cities]
      summary: Search cities
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/City' }
...
  /users/search:
    get:
      tags: [Users]
      summary: Search users
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
...
  /users/{userId}/report:
    post:
      tags: [Users]
      summary: Report a legacy user
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200':
          description: Report accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessMessage' }
  /search:
    get:
      tags: [Search]
      summary: Global search
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: type
          in: query
          schema: { type: string, enum: [all, events, cities, restaurants, users, tags, categories] }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SearchResponse' }
...
  /admin/users:
    get:
      tags: [Admin]
      summary: List account users
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of account users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AccountUser' }
...
  /admin/events:
    get:
      tags: [Admin]
      summary: List events
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Event' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success: { type: boolean }
        message: { type: string }
...
    Event:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        category_id: { type: integer }
        restaurant_id: { type: integer }
        current_participants: { type: integer }
        max_participants: { type: integer }
        price: { type: number }
        date: { type: string }
        start_time: { type: string }
        end_time: { type: string }
    Restaurant:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        city: { type: string }
...
    Tag:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        category_id: { type: integer }
    City:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        latitude: { type: number }
        longitude: { type: number }
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        alias: { type: string }
        bio: { type: string }
        profile_picture_url: { type: string }
        banner_picture_url: { type: string }
    AccountUser:
      type: object
      properties:
        id: { type: integer }
        clerk_user_id: { type: string }
        email: { type: string }
        first_name: { type: string }
        last_name: { type: string }
        role: { type: string }
    SearchResponse:
      type: object
      properties:
        type: { type: string }
        results:
          type: object
...